+++ src/config.rs
@@ -21,6 +21,7 @@ mod v0_1;
 mod v0_2_dev_0;
 mod v0_2_dev_1;
 mod v0_2_dev_2;
+mod v0_2_dev_3;
 
 // NOTE: When you switch to a new version:
 //
@@ -34,7 +35,7 @@ mod v0_2_dev_2;
 // - write an updater in `ConfigUpdater`,
 // - update the previous updaters as well,
 // - update `git z update`.
-pub use v0_2_dev_2::{Config, Scopes, Templates, Ticket};
+pub use v0_2_dev_3::{Config, Scopes, Templates, Ticket};
 
 use std::{fs, io, path::PathBuf, process::Command};
 
@@ -94,7 +95,7 @@ struct MinimalConfig {
 /// The name of the configuration file.
 pub const CONFIG_FILE_NAME: &str = "git-z.toml";
 /// The current version of the configuration file.
-pub const VERSION: &str = "0.2-dev.2";
+pub const VERSION: &str = "0.2-dev.3";
 
 const DEFAULT_TEMPLATE: &str = include_str!("../templates/COMMIT_EDITMSG");
 
@@ -109,7 +110,7 @@ impl Default for Config {
                 ),
                 (String::from("fix"), String::from("patches a bug")),
             ]),
-            scopes: None,
+            scopes: Some(Scopes::Any),
             ticket: None,
             templates: Templates {
                 commit: String::from(DEFAULT_TEMPLATE),
@@ -137,6 +138,10 @@ impl Config {
 
         match minimal_config.version.as_str() {
             VERSION => Ok(toml::from_str(toml)?),
+            "0.2-dev.2" => {
+                let config: v0_2_dev_2::Config = toml::from_str(toml)?;
+                Ok(config.into())
+            }
             "0.2-dev.1" => {
                 let config: v0_2_dev_1::Config = toml::from_str(toml)?;
                 Ok(config.into())
@@ -171,6 +176,41 @@ fn repo_root() -> Result<PathBuf, RepoRootError> {
     } else {
         let git_error = String::from_utf8(git_rev_parse.stderr)?;
         Err(RepoRootError::GitError(git_error.trim().to_owned()))
+    }
+}
+
+impl From<v0_2_dev_2::Config> for Config {
+    fn from(old: v0_2_dev_2::Config) -> Self {
+        Self {
+            version: old.version,
+            types: old.types,
+            scopes: old.scopes.map(Into::into),
+            ticket: old.ticket.map(Into::into),
+            templates: old.templates.into(),
+        }
+    }
+}
+
+impl From<v0_2_dev_2::Scopes> for Scopes {
+    fn from(old: v0_2_dev_2::Scopes) -> Self {
+        match old {
+            v0_2_dev_2::Scopes::List { list } => Self::List { list },
+        }
+    }
+}
+
+impl From<v0_2_dev_2::Ticket> for Ticket {
+    fn from(old: v0_2_dev_2::Ticket) -> Self {
+        Self {
+            required: old.required,
+            prefixes: old.prefixes,
+        }
+    }
+}
+
+impl From<v0_2_dev_2::Templates> for Templates {
+    fn from(old: v0_2_dev_2::Templates) -> Self {
+        Self { commit: old.commit }
     }
 }